pyproject.toml
[project]
name = "mcp-sqlalchemy-server"
version = "0.1.0"
description = "A simple MCP ODBC server using FastAPI and ODBC"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
 "mcp[cli]>=1.4.1",
 "pyodbc>=5.2.0",
 "python-dotenv>=1.0.1",
 "setuptools>=77.0.1",
 "sqlalchemy>=2.0.36",
 "virtuoso-sqlalchemy",
]
[[project.authors]]
name = "Sergey Malinin"
email = "sergmalinin@gmail.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[tool.uv.sources]
virtuoso-sqlalchemy = { git = "https://github.com/OpenLinkSoftware/Virtuoso-SQLAlchemy.git" }

[project.scripts]
mcp-server-odbc = "mcp_sqlalchemy_server:main"
